requirements.txt
numpy==1.26.2
pandas==2.1.4
matplotlib==3.4.3
flask==3.0.0
seaborn==0.11.2
Pillow==10.1.0
opencv-python==4.8.1.78
mysql-connector-python==8.2.0
gunicorn==21.2.0
